SanskritiAR_Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic[email]
python-dotenv
google-generativeai
firebase-admin